[sys-kernel/linux-sabayon-sources-2.6.32] Updated ebuild for compatibility.
Updated sys-kernel/linux-sabayon-sources-2.6.32 to provide compatiblity
with the sys-kernel/linux-sabayon package, and provided an additional
USE flag (sources_standalone) to toggle included files to provide a
working set of kernel sources that doesn't have file collisions with the
sys-kernel/linux-sabayon package.  The default for the
"sources_standalone" USE flage is disabled, which will provide a
dependency on sys-kernel/linux-sabayon-2.6.32.  This is the desired
behavior for an entropy build of this package.