[www-client/mozilla-firefox] Updates for default configuration.
The change provides a default bookmarks file that matches the
default bookmarks file used by Sabayon.  It also fixes a bug in
the firefox-default-prefs.js file that was preventing incorporation
of all the settings due to a spelling error.  Some entries (which
weren't working due to the bug anyways) were removed since they
weren't working as intended once they were allowed to work.