[lxde-base/lxdm] Added new ebuild for lxde-base/lxdm-0.2.0
While Gentoo is supporting a good number of lxde ebuilds,
lxdm is currently lightly supported due to it's expirmental status.
The lxde-base/lxdm-0.2.0 package currently has some shortcomings.
It will not work correctly when configured in /etc/conf.d/xdm
since it causes /etc/init.d/xdm to hang due to the program not
exiting like a daemon.

This ebuild is being provided for evaluation, but needs further
development in the scripts that launch it.