[dev-libs/dbus-c++] Removed since net-p2p/linkage was removed.
The only package that depended on dev-libs/dbus-c++ was net-p2p/linkage,
which was removed since the upstream went dead.  This ebuild was
not being maintained, and was not tracking the change in the upstream
source location to http://sourceforge.net/projects/dbus-cplusplus/

It can be re-added if anything ever needs it, and rebuilt to
reflect the upstream changes.