[sys-apps/openrc] Fix handling of clockskew on shutdown.
Openrc scans for clock skew on system start-up to determine if it
should save the dependency cache on shutdown.  If a clockskew is
detected, the dependency cache is not saved.

However, /etc/init.d/savecache would return a "1" error code,
which halted further processing of shutdown scripts, and causes
the computer to shutdown unclean.

Since the calling program does not know how to handle an error
due to clock skew, the /etc/init.d/savecache script returns a
"0" now, to allow the computer to finish unmounting drives and
perform a clean shutdown.