[app-misc/kupfer] Adding ebuild for app-misc/kupfer.
Adding an ebuild to the overlay for app-misc/kupfer to address
Sabayon bug support request:
http://bugs.sabayon.org/show_bug.cgi?id=1449
The ebuild was sourced from Gentoo bugzilla link:
http://bugs.gentoo.org/show_bug.cgi?id=307187
and from the keruspe overlay.
The ebuild, as it stands with this commit, has a serious Sandbox
violation due to setting mime types before the file locations
have been finalized.