[kde-base/oxygen-icons] Fixed USE Flag conflict.
Newer versions of portage were complaining (without really telling
you why) if the kde-base/oxygen-icons does not include a 'aqua'
USE Flag.  This may be buried somewhere in the eclass, but this is
a drawback of intricate eclasses for managing ebuilds.

Portage seems to be perfectly fine building kde-base/oxygen-icons with
"-aqua", but it damn well better be there apparently.