[www-plugins/google-talkplugin] Adding new ebuild for Google Talk.
Adding a new ebuild for a Linux version of Google Talk.
This is a binary file.  The source code is not available at this time.
Note the license is listed as "Unknown".  Google has not released
this binary file without clarifing the license.

This ebuild was based on google-talkplugin-1.4.1.0.1-r100 from
http://gpo.zugaina.org/www-plugins/google-talkplugin