[x11-plugins/desklet-gauges] Updated new source location.
I've updated the source location and homepage for this ebuild.
Apparently, the original gdesklets project died.  Some people
have been trying to pick up the project.  This particular package
is not being maintained at the moment, and is a archived copy
kept at http://www.gdesklets.info/archive/.

I've confirmed the check sums for the tarball stored at this
location match the original check sums.

For now, the package continues to build OK.  If bugs develop,
we'll have to decide to adopt the package, or kick it to the curb.

I've also marked the package as "~amd64" based on James Cook's
feedback that it builds OK at least on his installation.