[sys-fs/btrfs-progs] Update btrfs-progs-0.19 with patchset
The official btrfs-progs release has become out-of-date.  This
issue is acknowledged by the btrfs developers, and they plan to
issue an updated version when they clear a few more pressing problems
away.

In the meantime, the btrfs user community has identified the candidate
patch set for the next release of btrfs-progs.

These can be found at:
http://git.darksatanic.net/cgi/gitweb.cgi?p=btrfs-progs-unstable.git;a=summary

This patch set is based on the 'for-chris' branch of this repository.