[gnome-base/gsettings-desktop-schemas] port from tree, set show-desktop-icons=true by default
The override config files support in GNOME3 is still buggy and show-desktop-icons
doesn't seem to work correctly during the first GNOME3 startup.
This ebuild includes a patch to enable the setting by default, hopefully
workarounding the issue. We need to show the desktop icons in order
to make possible to start the Installer on the Live system.