[dev-python/pygobject*] provide split ebuilds version of pygobject
dev-python/pygobject[cairo] is starting to be required for several
graphical application that otherwise fail to run at runtime. The
list of apps includes the in-development Entropy Rigo Browser.
Unfortunately, dev-python/pygobject[cairo] pulls in a shitload
of deps: half X -> mesa -> llvm/clang. This is unwanted on
X-less images we have (SpinBase) due to the big amount of extra
software we would have to ship with.
We cannot remove dev-python/pygobject from ISO images because
Anaconda requires it.