[sys-devel/base-gcc] Clean-up Partial EAPI Change
When implementing a check for a valid gcc profile to build the
split Sabayon gcc packages, I wanted to place that check into
the package_pretend() phase, which required an EAPI change.

I later discovered that the toolchain eclass is not structured to
fully support higher EAPI's, but I did not fully revert my changes.