[x11-drivers/xf86-video-ati] fork, in order to enable modesetting via modprobe.d file
Sabayon default kernels ship with KMS disabled by default to
allow fglrx to work. However, latest radeon drivers require KMS
to be enabled in order to work properly. For this reason, one
solution could be to enable KMS via modprobe.d/.

However, it is still unclear how our Live systems behave given this
change. But we will find out with a bit of trial and error.