[x11-drivers/ati-drivers] rename USE=multilib to USE=x-multilib
Sabayon uses a multilib profile by default and it turns out that it's
impossible to disable USE=multilib on a per-package basis.
Given the huge amount of libraries pulled in by USE=multilib for both
ati-userspace and nvidia-userspace, it has been decided to disable the
installation of /usr/lib32 stuff for these two specific packages.
To workaround the Portage limitations, USE=multilib has been renamed to
USE=x-multilib.