[sys-devel/base-gcc] Fix error with 'go' USE flag.
The upstream toolchain eclass has changed so that the 'go'
USE flag is no longer in IUSE for <gcc-4.7.

Portage throws out an error now if flags are not in IUSE.

The only place that the 'go' USE flag is implemented in this
ebuild is for a PDEPEND on sys-devel/gcc-config, however,
the upstream toolchain eclass alredy has a higher dependency,
and we already have the appropriate sys-devel/gcc-config
in Entropy.

So I just removed the PDEPEND line, deferring to the PDEPEND
in the toolchain eclass.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>