[gnome-extra/zenity] fork ebuild to fix bug 4479
fixes:
zenity 3.8 --list outputs first answer twice separated by a pipe

The only (functional) difference is the added patch.
It should be fixed in 3.10.

This bug may be related to the Gtk+ version (this is a guess),
and given that Sabayon has the new one (but old zenity), this bug
may be Sabayon specific. That's why the ebuild is in this overlay.