[dev-python/qscintilla-python-qt*,x11-libs/qscintilla-qt*] new split ebuilds
Note that the ebuilds are copied from Gentoo and not changed in any way!
Configuration has to be contolled using USE flags (I checked that it
works and makes sense for these packages.)

Note: qscintilla (and -python) 2.10 allows qt4 with qt5, unlike the
previous version which allowed for only one of them.

Split ebuilds introduced to avoid bloat.