[x11-misc/sddm] fix permissions under autologin
On automatic start, the user didn't have permissions to configure wifi
etc. It appears that root's session from /dev/tty1 was being in effect.
Changing VT in sddm.conf to some other than the first one fixes it, so
does (re-)adding the Conflicts line.
(The second way is more in line with default setting so I did that.)