[dev-vcs/git*,www-apps/gitweb] version bump to 2.17.0
The first version bump with the new split ebuilds, that use a post
cleanup script at the end of src_install.

Thanks to this, there were only a few conflicts per ebuild (when merged
from Gentoo 2.17.0 on top of 2.16.2) and conflict resolution was
strightforward: related to KEYWORDS, PN -> MY_PN and the block that was
changed in gitweb.

Updating the "spec" file isn't the most friendly thing ever but it could
be advanced further as needed.