[net-dns] add Jinja2 template and gen. script for avahi
It generates exactly how the ebuilds look like for easier testing.
As stated in the comment, it can be cleaned a bit, but still it could be
easier during version and revision bumps (applying changes to ebuilds.)
Time will tell; it's an experiment, and moving back to previous
(hopefully only after due consideration!) is as easy as just starting
to use the generated ebuilds.