[net-analyzer/net-snmp] fork ebuild, disable wrong checks
Gentoo 679880. Checks on the build host for /dev/kmem (and more) and
hard codes related information in the binary.

If /dev/kmem was on the server but is not on the client side, there is
an error: Agent initialization failed.

The only change is the added patch.

Thanks to optiz0r for reporting and checking stuff!