[app-portage/unsymlink-lib] switch dep. to python:2.7
On system that doesn't have Python 3.6 already, it is installed early to
satisfy unsymlink-lib's dependency... and it borks because the
installation is done before the profile migration.

I removed the python:3.6 dependencies to be on the safer side with
respect to package installation order.

Credits for noticing go to some optiz0r's VM. Thanks, VM! ;)