[app-portage/unsymlink-lib] remove all dependencies
Hopefully it will help Entropy consider it first. It still has to be
seen if it's enough. Now, some other packages can be upgraded first but
that's not good.

Something like priorities in packages.db.critical would be better but
even Entropy cannot be always upgraded before the profile migration
completes safely, so patching the code to do /things/ is out of question.

Thanks to dyingmuppet for testing and chat where this idea was born.